+++ modules/function_handle/builtin/include/str2funcBuiltin.hpp
@@ -10,12 +10,13 @@
 #pragma once
 //=============================================================================
 #include "ArrayOf.hpp"
+#include "Evaluator.hpp"
 //=============================================================================
 namespace Nelson {
 //=============================================================================
 namespace FunctionHandleGateway {
     ArrayOfVector
-    str2funcBuiltin(int nLhs, const ArrayOfVector& argIn);
+    str2funcBuiltin(Evaluator* eval, int nLhs, const ArrayOfVector& argIn);
 }
 //=============================================================================
 } // namespace Nelson