+++ CHANGELOG.md
@@ -10,6 +10,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 - localization files `.pot`, `.po` updated.
 
+### Added
+
+- `str2func` manages anonymous function.
+
 ## 0.6.6 (2022-06-27)
 
 ### Added